Fix queries with ByteStrings
Problem: when debug logging is enabled, `decodeUtf8` is applied to the
query string prior to putting it to the logger. But decoding to UTF-8
can fail, which will cause failure of entire query.

Solution: use lenient version of decoding. We don't care much about
preserving the exact original query in logging.